Force transaction rollback on query execution errors
Basically, any queries that modify the database need to be wrapped
in a try/except block so that if an exception is hit (and realistically
it only needs to be the execute() call), we set the rollback attribute
on SqlTool to True. This will then cause that SqlTool to issue a rollback
when it falls out of scope.