Add support for folders
This is really two related changes:
1. Add a 'folder' value to the doctype enum and remove 'isfolder' from
the Documents table.
2. Allow for the creation of folders and for documents to subsequently
be set as a child of those folders. We use Common Table Subexpressions
(or CTEs or "recursive queries") to recursively build a list of folders
in a course, inserting spaces when necessary to denote the hierarchy in a folder
list.

Subsequent updates will add support for viewing those folders to the main
document view.